+++ src/icons.ts
@@ -1,4 +1,7 @@
-import mappingsJSON from '../mappings.json' with {type: 'json'};
+import {createRequire} from 'node:module';
+
+const require = createRequire(import.meta.url);
+const mappingsJson = require('../mappings.json') as typeof import('../mappings.json'); // eslint-disable-line @typescript-eslint/consistent-type-imports
 
 /**
  * Contains Nerd Fonts icon.
@@ -30,11 +33,11 @@ export type Mapping = {
 /**
  * Contains all Nerd Fonts icons.
  */
-export const icons = Object.fromEntries(Object.entries(mappingsJSON).map(
+export const icons = Object.fromEntries(Object.entries(mappingsJson).map(
 	([classname, charHex]) => ([classname, {
 		char: String.fromCodePoint(charHex),
 		hexCode: charHex,
 	} satisfies NerdIcon]),
-)) as Record<keyof typeof mappingsJSON, NerdIcon>;
+)) as Record<keyof typeof mappingsJson, NerdIcon>;
 
 export default icons;