+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Use createRequire instead of json esm import.
+- Update the README.
+- Change the description.
+
 ## [0.1.1] - 2024-09-19
 
 - Add extension iterations: '.d.ts' -> '.ts'. Seti mappings fix.